Enum directx_sys::dwrite::FontFeatureTag

pub enum FontFeatureTag {
    AlternativeFractions,
    PetiteCapitalsFromCapitals,
    SmallCapitalsFromCapitals,
    ContextualAlternates,
    CaseSensitiveForms,
    GlyphCompositionDecomposition,
    ContextualLigatures,
    CapitalSpacing,
    ContextualSwash,
    CursivePositioning,
    Default,
    DiscretionaryLigatures,
    ExpertForms,
    Fractions,
    FullWidth,
    HalfForms,
    HalantForms,
    AlternateHalfWidth,
    HistoricalForms,
    HorizontalKanaAlternates,
    HistoricalLigatures,
    HalfWidth,
    HojoKanjiForms,
    JIS04Forms,
    JIS78Forms,
    JIS83Forms,
    JIS90Forms,
    Kerning,
    StandardLigatures,
    LiningFigures,
    LocalizedForms,
    MarkPositioning,
    MathematicalGreek,
    MarkToMarkPositioning,
    AlternateAnnotationForms,
    NLCKanjiForms,
    OldStyleFigures,
    Ordinals,
    ProportionalAlternateWidth,
    PetiteCapitals,
    ProportionalFigures,
    ProportionalWidths,
    QuarterWidths,
    RequiredLigatures,
    RubyNotationForms,
    StylisticAlternates,
    ScientificInferiors,
    SmallCapitals,
    SimplifiedForms,
    StylisticSet1,
    StylisticSet2,
    StylisticSet3,
    StylisticSet4,
    StylisticSet5,
    StylisticSet6,
    StylisticSet7,
    StylisticSet8,
    StylisticSet9,
    StylisticSet10,
    StylisticSet11,
    StylisticSet12,
    StylisticSet13,
    StylisticSet14,
    StylisticSet15,
    StylisticSet16,
    StylisticSet17,
    StylisticSet18,
    StylisticSet19,
    StylisticSet20,
    Subscript,
    Superscript,
    Swash,
    Titling,
    TraditionalNameForms,
    TabularFigures,
    TraditionalForms,
    ThirdWidths,
    Unicase,
    VerticalWriting,
    VerticalAlternatesAndRotation,
    SlashedZero,
}

Variants

AlternativeFractions

PetiteCapitalsFromCapitals

SmallCapitalsFromCapitals

ContextualAlternates

CaseSensitiveForms

GlyphCompositionDecomposition

ContextualLigatures

CapitalSpacing

ContextualSwash

CursivePositioning

Default

DiscretionaryLigatures

ExpertForms

Fractions

FullWidth

HalfForms

HalantForms

AlternateHalfWidth

HistoricalForms

HorizontalKanaAlternates

HistoricalLigatures

HalfWidth

HojoKanjiForms

JIS04Forms

JIS78Forms

JIS83Forms

JIS90Forms

Kerning

StandardLigatures

LiningFigures

LocalizedForms

MarkPositioning

MathematicalGreek

MarkToMarkPositioning

AlternateAnnotationForms

NLCKanjiForms

OldStyleFigures

Ordinals

ProportionalAlternateWidth

PetiteCapitals

ProportionalFigures

ProportionalWidths

QuarterWidths

RequiredLigatures

RubyNotationForms

StylisticAlternates

ScientificInferiors

SmallCapitals

SimplifiedForms

StylisticSet1

StylisticSet2

StylisticSet3

StylisticSet4

StylisticSet5

StylisticSet6

StylisticSet7

StylisticSet8

StylisticSet9

StylisticSet10

StylisticSet11

StylisticSet12

StylisticSet13

StylisticSet14

StylisticSet15

StylisticSet16

StylisticSet17

StylisticSet18

StylisticSet19

StylisticSet20

Subscript

Superscript

Swash

Titling

TraditionalNameForms

TabularFigures

TraditionalForms

ThirdWidths

Unicase

VerticalWriting

VerticalAlternatesAndRotation

SlashedZero

Methods

impl FontFeatureTag

unsafe fn from_chars(tag: &[u8; 4]) -> FontFeatureTag

Create a custom OpenType tag from a 4-character byte string.

Marked as unsafe as it allows you to create discriminants that are not part of the enum definition which is undefined behaviour.

use directx_sys::dwrite::FontFeatureTag;
let rust_tag = unsafe { FontFeatureTag::from_chars(b"rust") };

fn to_chars(&self) -> [u8; 4]